fix: move constants to same file
